requirements.txt
flask
pyserial
Flask-SQLAlchemy
Flask-SocketIO
requests
eventlet